Refactor- mid:
this idea is being canceled, this branch is useless
